// src/app/dashboard/modules/agency-settings/enum/account-setting.enum.ts
export enum ESortOrder {
  default = 'default',
  descending = 'DESC',
  ascending = 'ASC'
}

export enum EPolicyTableColumnName {
  name = 'name',
  createdAt = 'createdAt',
  updatedAt = 'updatedAt',
  questions = 'questions'
}

export enum EPolicyType {
  create = 'create',
  update = 'update',
  delete = 'delete'
}